refactor: search바 공통컴포넌트 분리
