+++ src/components/common/Select.tsx
@@ -1,10 +1,12 @@
 import styled from "styled-components";
 
 const SelectStyle = styled.select`
-  width: 100%;
+  width: 80px;
+  height: 30px;
   max-width: 90px;
   padding: 6px;
   font-size: 12px;
+  outline: none;
 `;
 
 const Select = () => {
