+++ src/pages/Catcher.tsx
@@ -1,5 +1,4 @@
 import PlayerCard from "@components/Player/PlayerCard";
-import SearchForm from "@components/SearchForm";
 import { PlayerContentsWrapper } from "@styles/PlayerTable.style";
 import { useState } from "react";
 
@@ -100,7 +99,6 @@ const Catcher = () => {
   const [playerList, setPlayerList] = useState(dataList);
   return (
     <>
-      <SearchForm />
       <PlayerContentsWrapper>
         <ul>
           {playerList.map((data) => (
