+++ src/pages/Coach.tsx
@@ -1,5 +1,4 @@
 import PlayerCard from "@components/Player/PlayerCard";
-import SearchForm from "@components/SearchForm";
 import { PlayerContentsWrapper } from "@styles/PlayerTable.style";
 import { useState } from "react";
 
@@ -97,7 +96,6 @@ const Coach = () => {
 
   return (
     <>
-      <SearchForm />
       <PlayerContentsWrapper>
         <ul>
           {playerList.map((data) => (
